Crate cumulus_client_consensus_relay_chain

The relay-chain provided consensus algoritm for parachains.

This is the simplest consensus algorithm you can use when developing a parachain. It is a permission-less consensus algorithm that doesn’t require any staking or similar to join as a collator. In this algorithm the consensus is provided by the relay-chain. This works in the following way.

Each node that sees itself as a collator is free to build a parachain candidate.

This parachain candidate is send to the parachain validators that are part of the relay chain.

The parachain validators validate at most X different parachain candidates, where X is the total number of parachain validators.

The parachain candidate that is backed by the most validators is choosen by the relay-chain block producer to be added as backed candidate on chain.

After the parachain candidate got backed and included, all collators start at 1.

Structs

BuildRelayChainConsensusParams

Paramaters of build_relay_chain_consensus.

RelayChainConsensus

The implementation of the relay-chain provided consensus for parachains.

Verifier

A verifier that just checks the inherents.

Functions

build_relay_chain_consensus

Build the RelayChainConsensus.

import_queue

Start an import queue for a Cumulus collator that does not uses any special authoring logic.